Add tests and refactor exising array method tests into three separate files
